Add new tool swap and restock hooks and change some of that stuff
The tool swap and restock should now catch hopefully all of the left and
right clicks and item uses.

However there is an issue with the inventory sync from the server to the
client... If the client is spamming use packets, then the synced items
from the server will kind of reset the durability of the item so that many
of the uses are lost from the client's apparent tool durability, and thus
the client may overuse and break the item due to this. I can't think of
any proper non-super-hacky solution to this problem...