Add a hack fix workaround for inventory sync issues with fast clickers
The items in the main hand and offhand slot are not allowed to be synced
from the server while the fast clicking tweaks are active and the button
held down. This seems to fix issues with the hand restock and tool swap
features, where the server previously would sync an outdated state of the
stack to the client, and thus the mod would lose track of the actual stack
size or durability left.