Add a Placement Grid restriction tweak (only works with Fast Placement atm!)
The issue with it is targeted position vs. placement position stuff,
that's why it currently only works with Fast Placement mode enabled, as it
offsets the position. This needs to be sorted when the entire placement
tweaks stuff gets rewritten/cleaned up at some point

The issue with it is targeted position vs. placement position stuff,
that's why it currently only works with Fast Placement mode enabled, as it
offsets the position. This needs to be sorted when the entire placement
tweaks stuff gets rewritten/cleaned up at some point..